GalleryのUIのリファクタリング。
